added an info/welcome screen
