+++ src/client/component/NavLink.jsx
@@ -1,0 +1,8 @@
+import React from 'react'
+import { Link } from 'react-router'
+
+export default React.createClass({
+    render() {
+        return <Link {...this.props} activeClassName="active"/>
+    }
+})
