+++ src/client/component/pages/Home.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+
+export default React.createClass({
+    render() {
+        return <div>
+            <h3>Willkommen bei der Stupa Wahl 2017</h3>
+            <p>
+                Info geblubber...
+            </p>
+        </div>
+    }
+})
