+++ src/client/style/main.scss
@@ -1,1 +1,5 @@
 @import 'reset';
+
+.active {
+  color: #00FF00;
+}
